Improved compile time and class function instantiation.

+++ src/caffe/libdnn/libdnn_conv.cpp
@@ -1548,10 +1548,12 @@ void LibDNNConv<MItype, MOtype>::Forward(vptr<const MItype> bottom_data,
   int fw_div_M = fw_wptm * fw_wgs1;
 
   shared_ptr<DeviceKernel> kernel = this->program_->GetKernel("conv_forward");
-  vector<size_t> group = {((this->N_FW_ - 1) / fw_div_N + 1),
-                          ((this->M_FW_ - 1) / fw_div_M + 1),
-                          batch_size * group_};
-  vector<size_t> local = {fw_wgs0, fw_wgs1, 1};
+  vector<size_t> group = {static_cast<size_t>(
+                            ((this->N_FW_ - 1) / fw_div_N + 1)),
+                          static_cast<size_t>((this->M_FW_ - 1) / fw_div_M + 1),
+                          static_cast<size_t>(batch_size * group_)};
+  vector<size_t> local = {static_cast<size_t>(fw_wgs0),
+                          static_cast<size_t>(fw_wgs1), 1};
 
   if (bias_term_) {
     kernel->add_arg(&bottom_data);
@@ -1601,10 +1603,13 @@ void LibDNNConv<MItype, MOtype>::Backward(bool prop_down_data,
   if (prop_down_data) {
     shared_ptr<DeviceKernel> kernel =
         this->program_->GetKernel("conv_backward");
-    vector<size_t> group = {((this->N_BW_ - 1) / bw_div_N + 1),
-                            ((this->M_BW_ - 1) / bw_div_M + 1),
-                            batch_size * group_};
-    vector<size_t> local = {bw_wgs0, bw_wgs1, 1};
+    vector<size_t> group = {static_cast<size_t>(
+                              ((this->N_BW_ - 1) / bw_div_N + 1)),
+                            static_cast<size_t>(
+                              ((this->M_BW_ - 1) / bw_div_M + 1)),
+                            static_cast<size_t>(batch_size * group_)};
+    vector<size_t> local = {static_cast<size_t>(bw_wgs0),
+                            static_cast<size_t>(bw_wgs1), 1};
 
     if (bias_term_) {
       kernel->add_arg(&top_diff);
@@ -1624,10 +1629,13 @@ void LibDNNConv<MItype, MOtype>::Backward(bool prop_down_data,
   if (prop_down_weights && (this->weights_backward_ || this->bias_backward_)) {
     shared_ptr<DeviceKernel> kernel =
         this->program_->GetKernel("conv_weights");
-    vector<size_t> group = {((this->N_WG_ - 1) / wg_div_N + 1),
-                            ((this->M_WG_ - 1) / wg_div_M + 1),
+    vector<size_t> group = {static_cast<size_t>(
+                              ((this->N_WG_ - 1) / wg_div_N + 1)),
+                            static_cast<size_t>(
+                              ((this->M_WG_ - 1) / wg_div_M + 1)),
                             0};
-    vector<size_t> local = {wg_wgs0, wg_wgs1, 1};
+    vector<size_t> local = {static_cast<size_t>(wg_wgs0),
+                            static_cast<size_t>(wg_wgs1), 1};
     if (wgalgo_ == LIBDNN_CONVOLUTION_WG_ALGO_DIRECT) {
       group[2] = group_;
     }
